[LIH] Reuse same mailchimp form across entire site
- Make `full_name` an optional field.
- Move mailchimp AJAX form handling into `home.js` so it runs on all
  pages, not just the Area page.
- Tweak mailchimp form JS so that errors are only displayed on the form
  you’ve interacted with, if the same form appears multiple times on the
  page.
