+++ hub/static/js/area.js
@@ -2,7 +2,6 @@ import $ from 'jquery/dist/jquery.slim'
 import { Chart, BarController, BarElement, CategoryScale, LinearScale, Legend, Tooltip } from 'chart.js'
 import ChartDataLabels from 'chartjs-plugin-datalabels';
 import trackEvent from './analytics.esm.js'
-import setUpCollapsable from './collapsable.esm.js'
 import Dropdown from 'bootstrap/js/dist/dropdown'
 
 
@@ -17,20 +16,6 @@ Chart.defaults.animation.duration = 0
 Chart.defaults.responsive = true
 
 import setUpAreaPage from './area.esm.js'
-
-async function mailingListSignup($form) {
-    const response = await fetch($form.attr('action'), {
-        method: $form.attr('method') || 'GET',
-        mode: 'cors',
-        credentials: 'same-origin',
-        body: $form.serialize(),
-        headers: {
-            "Content-Type": 'application/x-www-form-urlencoded',
-            "Accept": 'application/json; charset=utf-8',
-        },
-    })
-    return response.json()
-}
 
 $(function(){
     if( 'geolocation' in navigator ) {
@@ -112,46 +97,6 @@ $(function(){
         })
     })
 
-    $('.js-collapsable-mailing-list-form').each(function(){
-        setUpCollapsable(
-            $(this).find('.js-mailing-list-name, .js-mailing-list-extras'),
-            $(this).find('.js-mailing-list-email input#email'),
-            'keyup change',
-            function($targets, $triggers){
-                return $triggers.eq(0).val() !== '';
-            }
-        );
-    });
-
-    $('.js-mailing-list-signup').on('submit', function(e){
-        e.preventDefault();
-        var $form = $(this);
-        $('.invalid-feedback').remove()
-        mailingListSignup($form).then(function(response){
-            if (response['response'] == 'ok') {
-                $form.hide()
-                $('.js-mailing-list-success').removeClass('d-none')
-            } else {
-                console.log(response)
-                for (var k in response["errors"]) {
-                    var id = '#' + k
-                    var el = $(id)
-                    el.addClass('is-invalid')
-                    var error_el = $('<div>')
-                    error_el.addClass('invalid-feedback d-block fs-6 mt-2')
-                    error_el.html( '<p>' + response["errors"][k].join(", ") + '</p>' )
-                    el.after(error_el)
-                }
-
-                if ("mailchimp" in response["errors"]) {
-                    var error_el = $('<div>')
-                    error_el.addClass('invalid-feedback d-block fs-6 mt-2')
-                    error_el.html( '<p>There was a problem signing you up, please try again.</p>' )
-                    $form.before(error_el)
-                }
-            }
-        });
-    })
 })
 
 var makeChart = function() {
