+++ hub/templates/hub/area.html
@@ -855,7 +855,7 @@
             </div>
           {% endif %}
             <div class="col-lg-6">
-                {% include 'hub/includes/mailing-list-form.html' with classes="js-collapsable-mailing-list-form" %}
+                {% include 'hub/includes/mailing-list-form.html' with classes="js-collapsable-mailing-list-form" id_prefix="area_" %}
             </div>
         </div>
     </div>
