+++ hub/templates/hub/sign_up.html
@@ -5,7 +5,7 @@
 <div class="py-4 py-lg-5">
     <div class="container">
 
-        {% include 'hub/includes/mailing-list-form.html' %}
+        {% include 'hub/includes/mailing-list-form.html' with id_prefix="signup_" show_server_errors=True %}
 
     </div>
 </div>
