+++ hub/templates/hub/sources.html
@@ -15,7 +15,7 @@
         <div class="container">
             <div class="row">
                 <div class="col-md-5 col-xl-6">
-                    {% include 'hub/includes/mailing-list-form.html' with classes="js-collapsable-mailing-list-form" %}
+                    {% include 'hub/includes/mailing-list-form.html' with classes="js-collapsable-mailing-list-form" id_prefix="sources_" %}
                 </div>
                 <div class="col-md-5 mt-5 mt-md-0 mx-md-auto">
                     <h2 class="h4 mt-0">Got new data to contribute?</h2>
